+++ resources/views/posts/singlepost.blade.php
@@ -9,12 +9,15 @@
         {!!$post->body!!}
     </div>
     <hr>
-    <small>{{$post->created_at}}</small>
+    <small>{{$post->created_at}}  by {{$post->user->name}}</small>
     <hr>
+    @if(!Auth::guest())
+        @if(Auth::user()->id == $post->user_id)
+            <a href="/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
 
-    <a href="/posts/{{$post->id}}/edit" class="btn btn-primary">Edit</a>
-
-    {!! Form::open(['action' => ['App\Http\Controllers\PostsController@destroy',$post->id],'method'=>'DELETE','class'=>'float-right']) !!}
-    {{Form::submit('Delete',['class'=>'btn btn-danger'])}}
-    {!! Form::close() !!}
+            {!! Form::open(['action' => ['App\Http\Controllers\PostsController@destroy',$post->id],'method'=>'DELETE','class'=>'float-right']) !!}
+            {{Form::submit('Delete',['class'=>'btn btn-danger'])}}
+            {!! Form::close() !!}
+        @endif
+    @endif
 @endsection
